flask-backend/requirements.txt
Flask
flask-marshmallow
Flask-SQLAlchemy
marshmallow-sqlalchemy
markupsafe
psycopg2-binary
python-decouple
gunicorn